+++ src/Data/Aeson/Types.hs
@@ -29,6 +29,7 @@ module Data.Aeson.Types
     , typeMismatch
     , unexpected
     -- * Type conversion
+    -- ** Parsing
     , Parser
     , Result(..)
     , FromJSON(..)
@@ -37,12 +38,19 @@ module Data.Aeson.Types
     , parseEither
     , parseMaybe
     , parseFail
-    , ToJSON(..)
-    , KeyValue(..)
+    -- ** Parser error handling
     , modifyFailure
     , prependFailure
     , parserThrowError
     , parserCatchError
+    -- ** Parsing with paths
+    , IResult (..)
+    , ifromJSON
+    , iparse
+    , iparseEither
+    -- ** Encoding
+    , ToJSON(..)
+    , KeyValue(..)
 
     -- ** Keys for maps
     , ToJSONKey(..)